Remove redundant 'as'

+++ src/ui/mod.rs
@@ -795,7 +795,7 @@ impl<'a> TereTui<'a> {
         )?;
 
         let (w, h) = main_window_size()?;
-        let help_text = get_formatted_help_text(w as usize);
+        let help_text = get_formatted_help_text(w);
         for (i, line) in help_text.iter()
                 .skip(scroll)
                 .chain(vec![vec![]].iter().cycle()) // add empty lines at the end
